find by id

+++ domain/model/pokemon.go
@@ -1,11 +1,11 @@
 package model
 
 type Pokemon struct {
-	ID   uint
+	ID   uint64
 	Name string
 }
 
-func newPokemon(Id uint, name string) *Pokemon {
+func newPokemon(Id uint64, name string) *Pokemon {
 	p := Pokemon{Name: name, ID: Id}
 	return &p
 }
